Fix unsubscribed ACClient disconnects after 6 seconds
This issue was caused by the assumption that an ACClient would always have a subscription. However, when ACClient is connected but not subscribed, PingRoundWatcher.ping() was never called. This caused PingRoundWatcher to time out after 6 seconds, the pingLimit. The fix was just to call pingRoundWatcher.ping from ACClient instead of ACChannel.