Fix PingRoundWatcher attempted reconnect when unaunathorized
This change mirrors the javascrpipt changes found here: https://github.com/rails/rails/pull/34194

Rails 6's Action Cable added the "disconnect" message type which has "reason" and "reconnect" keys. Rails sets the "reconnect" value to false in situations such as when reject_unauthorized_connection is called.

The fix was to setFinish on the pingRoundWatcher when the new disconnect message type is received.